pyproject.toml
[project]
name = "hud_vf_gym"
version = "0.1.0"
description = "HUD environments for Reinforcement Learning through Verifiers"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "verifiers[all] @ git+https://github.com/jdchawla29/verifiers.git",
    "hud-python"
]

[build-system]
requires = ["hatchling", "editables>=0.5"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.5",
]

[tool.hatch.build]
include = ["hud_vf_gym/"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"